feat: 강수량 타입 변환 (int -> double)

+++ app/src/main/java/com/steelzoo/ownweather/ui/Mapper.kt
@@ -13,7 +13,7 @@ fun NowWeatherData.toWeatherDataUI():WeatherDataUI {
         skyState.toSkyStateUI(),
         temperature,
         humidity.toInt(),
-        oneHourPrecipitation.toInt(),
+        oneHourPrecipitation,
         windDirection,
         windSpeed,
         0,
